refactor: extract message and message-list logic into mixins
