+++ packages/message-list/package.json
@@ -37,6 +37,7 @@
     "polymer"
   ],
   "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
     "@polymer/polymer": "^3.0.0",
     "@vaadin/a11y-base": "24.6.0-beta1",
     "@vaadin/avatar": "24.6.0-beta1",
